requirements.txt
streamlit==1.21.0
tensorflow==2.17.0
pillow
numpy